requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
python-dateutil==2.8.0
python-pptx==0.6.18
pytz==2019.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.25.3
Werkzeug==0.15.4
XlsxWriter==1.1.8
